# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestión de Subvenciones API
  description: |
    API REST para el Sistema de Gestión de Subvenciones de Alquiler.
    
    ## Autenticación
    La API utiliza JWT (JSON Web Tokens) para autenticación. Los tokens se envían en el header `Authorization` con el prefijo `Bearer`.
    
    ## Roles
    - **BENEFICIARIO**: Solicitante de la subvención
    - **ARRENDADOR**: Propietario del inmueble
    - **DIGER**: Personal administrativo que gestiona solicitudes
    - **DIRECTORA**: Directora que aprueba solicitudes
    - **ORDENADOR_GASTO**: Responsable de firma final
    - **CRI**: Centro de Recursos Informáticos (auditoría)
    
    ## Estados del Proceso
    - BORRADOR → ENVIADA → EN_VALIDACION → PENDIENTE_DIRECTORA → APROBADA_DIRECTORA → REVISION_ORDENADOR → APROBADA → CERRADA
    - También: RECHAZADA, CORRECCION_SOLICITADA
  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: soporte@subvenciones.gob.es
  license:
    name: Propietario
servers:
  - url: http://localhost:4000/api
    description: Servidor de desarrollo
  - url: https://api.subvenciones.gob.es/api
    description: Servidor de producción
...
paths:
...
  /procesos:
    get:
      tags: [Procesos]
      summary: Listar procesos
      description: |
        Devuelve los procesos según el rol del usuario:
        - BENEFICIARIO: Solo sus propios procesos
        - ARRENDADOR: Procesos donde es arrendador
        - DIGER/DIRECTORA/ORDENADOR_GASTO: Todos los procesos
        - CRI: Todos los procesos (solo lectura)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: estado
          in: query
          schema:
            $ref: '#/components/schemas/EstadoProceso'
        - name: search
          in: query
          schema:
            type: string
            description: Buscar por código o nombre de beneficiario
      responses:
        '200':
          description: Lista de procesos
          content:
            application/json:
              schema:
                type: object
                properties:
                  procesos:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProcesoResumen'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer

    post:
      tags: [Procesos]
      summary: Crear proceso
      description: Solo rol DIGER puede crear procesos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [beneficiarioId]
              properties:
                beneficiarioId:
                  type: string
                  format: uuid
                arrendadorId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Proceso creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proceso'
        '403':
          $ref: '#/components/responses/Forbidden'
  /procesos/{id}:
    get:
      tags: [Procesos]
      summary: Obtener proceso por ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/procesoId'
      responses:
        '200':
          description: Proceso encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proceso'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /procesos/{id}/formulario:
    put:
      tags: [Procesos]
      summary: Actualizar formulario del proceso
      description: Solo en estado BORRADOR o CORRECCION_SOLICITADA
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/procesoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Datos del formulario (esquema libre)
      responses:
        '200':
          description: Formulario actualizado
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /procesos/{id}/decision:
    post:
      tags: [Procesos]
      summary: Registrar decisión sobre el proceso
      description: |
        Permite aprobar o rechazar según el rol:
        - DIGER: EN_VALIDACION → PENDIENTE_DIRECTORA o RECHAZADA
        - DIRECTORA: PENDIENTE_DIRECTORA → APROBADA_DIRECTORA o RECHAZADA
        - ORDENADOR_GASTO: REVISION_ORDENADOR → APROBADA o RECHAZADA
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/procesoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision:
                  type: string
                  enum: [APROBADO, RECHAZADO]
                observaciones:
                  type: string
      responses:
        '200':
          description: Decisión registrada
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /procesos/{id}/timeline:
    get:
      tags: [Procesos]
      summary: Obtener línea de tiempo del proceso
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/procesoId'
      responses:
        '200':
          description: Línea de tiempo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimelineEvent'

  # ==================== DOCUMENTOS ====================
...
  /documentos/{id}:
    get:
      tags: [Documentos]
      summary: Obtener metadatos del documento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Metadatos del documento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Documento'

    delete:
      tags: [Documentos]
      summary: Eliminar documento
      description: Solo en estado BORRADOR o CORRECCION_SOLICITADA
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Documento eliminado
        '400':
          description: No se puede eliminar en el estado actual
  /documentos/{id}/download:
    get:
      tags: [Documentos]
      summary: Descargar documento
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Archivo del documento
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /auditoria/procesos/{id}:
    get:
      tags: [Auditoría]
      summary: Obtener eventos de un proceso
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/procesoId'
      responses:
        '200':
          description: Eventos del proceso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventoAuditoria'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    procesoId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    Rol:
      type: string
      enum: [BENEFICIARIO, ARRENDADOR, DIGER, DIRECTORA, ORDENADOR_GASTO, CRI]
    EstadoUsuario:
      type: string
      enum: [ACTIVO, INACTIVO, BLOQUEADO]
    EstadoProceso:
      type: string
      enum:
        - BORRADOR
        - ENVIADA
        - EN_VALIDACION
        - CORRECCION_SOLICITADA
        - PENDIENTE_DIRECTORA
        - APROBADA_DIRECTORA
        - REVISION_ORDENADOR
        - APROBADA
        - RECHAZADA
        - CERRADA
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cedula:
          type: string
        email:
          type: string
          format: email
        nombreCompleto:
          type: string
        rol:
          $ref: '#/components/schemas/Rol'
        estado:
          $ref: '#/components/schemas/EstadoUsuario'
        mfaHabilitado:
          type: boolean
        creadoEn:
          type: string
          format: date-time
...
    ProcesoResumen:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codigo:
          type: string
          example: "SUB-2024-000001"
        estado:
          $ref: '#/components/schemas/EstadoProceso'
        beneficiario:
          type: object
          properties:
            id:
              type: string
            nombreCompleto:
              type: string
        creadoEn:
          type: string
          format: date-time
        actualizadoEn:
          type: string
          format: date-time
    Proceso:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codigo:
          type: string
        estado:
          $ref: '#/components/schemas/EstadoProceso'
        beneficiario:
          $ref: '#/components/schemas/User'
        arrendador:
          $ref: '#/components/schemas/User'
        formulario:
          type: object
        documentos:
          type: array
          items:
            $ref: '#/components/schemas/Documento'
        decisiones:
          type: array
          items:
            $ref: '#/components/schemas/Decision'
        creadoEn:
          type: string
          format: date-time
        actualizadoEn:
          type: string
          format: date-time
    Documento:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombreOriginal:
          type: string
        mimeType:
          type: string
        tamanoBytes:
          type: integer
        hash:
          type: string
        estado:
          type: string
          enum: [PENDIENTE, APROBADO, RECHAZADO]
        catalogo:
          $ref: '#/components/schemas/TipoDocumento'
        subidoPor:
          $ref: '#/components/schemas/User'
        creadoEn:
          type: string
          format: date-time
    TipoDocumento:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codigo:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        obligatorio:
          type: boolean
        extensionesPermitidas:
          type: array
          items:
            type: string
    Decision:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [APROBADO, RECHAZADO, CORRECCION_SOLICITADA]
        observaciones:
          type: string
        responsable:
          $ref: '#/components/schemas/User'
        fecha:
          type: string
          format: date-time
...
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
        descripcion:
          type: string
        responsable:
          type: object
          properties:
            nombre:
              type: string
            rol:
              $ref: '#/components/schemas/Rol'
        observaciones:
          type: string
        fecha:
          type: string
          format: date-time
    EventoAuditoria:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accion:
          type: string
        entidad:
          type: string
        entidadId:
          type: string
        detalles:
          type: object
        ip:
          type: string
        userAgent:
          type: string
        usuario:
          type: object
          properties:
            id:
              type: string
            nombreCompleto:
              type: string
        fecha:
          type: string
          format: date-time
security:
  - bearerAuth: []